+++ apps/sales/models.py
@@ -5,7 +5,7 @@ from apps.customers.models import Customer
 class Sale(models.Model):
     client = models.ForeignKey(Customer, on_delete=models.SET_NULL, null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True)
-    total = models.DecimalField(max_digits=12, decimal_places=2, default=0)
+    total = models.IntegerField(default=0)
     utilisateur = models.ForeignKey('auth.User', on_delete=models.SET_NULL, null=True, blank=True)
 
     class Meta:
@@ -19,8 +19,8 @@ class SaleItem(models.Model):
     vente = models.ForeignKey(Sale, on_delete=models.CASCADE, related_name='items')
     produit = models.ForeignKey(Product, on_delete=models.CASCADE)
     quantite = models.PositiveIntegerField()
-    prix_vente = models.DecimalField(max_digits=10, decimal_places=2)
-    sous_total = models.DecimalField(max_digits=10, decimal_places=2, editable=False)
+    prix_vente = models.IntegerField()
+    sous_total = models.IntegerField(editable=False)
 
     class Meta:
         verbose_name = "Ligne de vente"
